+++ lib/screen/add_menu/menu_image.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:flutter/material.dart';
+import 'package:image_picker/image_picker.dart';
 
 class MenuImage extends StatelessWidget {
   const MenuImage({Key? key, this.imagePath}) : super(key: key);
@@ -24,12 +25,17 @@ class MenuImage extends StatelessWidget {
         alignment: Alignment(1, -1),
         child: Padding(
           padding: EdgeInsets.all(8.0),
-          child: Icon(
-            Icons.change_circle,
-            size: 40,
-          ),
+          child: Icon(Icons.change_circle, size: 40),
+          // child: IconButton(
+          //     icon: Icon(Icons.change_circle), iconSize: 40, onPressed: () {}),
         ),
       )
     ]);
+  }
+
+  // TODO: 写真入れ替えメソッド
+  Future _getImage() async {
+    final ImagePicker _picker = ImagePicker();
+    final pickedFile = await _picker.pickImage(source: ImageSource.gallery);
   }
 }
