+++ lib/screen/main.dart
@@ -1,17 +1,12 @@
-import 'dart:io';
-
-import 'package:becos_kitchen/component/expandable_fab.dart';
-import 'package:becos_kitchen/screen/add_menu/add_menu.dart';
+import 'package:becos_kitchen/screen/menu_list/menu_list_page.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:image_picker/image_picker.dart';
-
-import '../component/menu_card.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp();
-  runApp(const MyApp());
+  runApp(const ProviderScope(child: MyApp()));
 }
 
 class MyApp extends StatelessWidget {
@@ -25,72 +20,7 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: const MyHomePage(),
+      home: const MenuListPage(),
     );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({Key? key}) : super(key: key);
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  final ImagePicker _picker = ImagePicker();
-  File? _image;
-
-  @override
-  Widget build(BuildContext context) {
-    var size = MediaQuery.of(context).size;
-    final double itemHeight = (size.height - kToolbarHeight) / 3;
-    final double itemWidth = size.width / 2;
-
-    return Scaffold(
-        appBar: AppBar(
-          title: const Text('メニュー一覧'),
-        ),
-        floatingActionButton: ExpandableFab(
-          distance: 80.0,
-          children: [
-            ActionButton(
-                onPressed: () => _getImage(ImageSource.camera),
-                icon: const Icon(Icons.add_a_photo)),
-            ActionButton(
-                onPressed: () => _getImage(ImageSource.gallery),
-                icon: const Icon(Icons.photo_album)),
-          ],
-        ),
-        body: Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: GridView.count(
-            childAspectRatio: (itemWidth / itemHeight),
-            crossAxisCount: 2,
-            children: [
-              MenuCard(),
-              MenuCard(),
-              MenuCard(),
-              MenuCard(),
-              MenuCard(),
-              MenuCard(),
-              MenuCard(),
-              MenuCard(),
-            ],
-          ),
-        ));
-  }
-
-  Future _getImage(ImageSource source) async {
-    final pickedFile = await _picker.pickImage(source: source);
-
-    setState(() {
-      if (pickedFile != null) {
-        _image = File(pickedFile.path);
-      }
-    });
-
-    Navigator.push(context,
-        MaterialPageRoute(builder: (context) => AddMenuScreen(image: _image)));
   }
 }
