+++ pubspec.yaml
@@ -36,7 +36,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   freezed_annotation: ^2.0.3
   json_annotation: ^4.5.0
-  riverpod: ^1.0.3
+  flutter_riverpod: ^1.0.3
   firebase_core: ^1.17.1
   flutter_tags: ^1.0.0-nullsafety.1
   flutter_rating_bar: ^4.0.1
